Move the CA sync to the vmc controller
